Color.wrap is the default when Color() is called.
It tries to parse the arguments into a color, and returns
white if it can't figure it out.

Color.wrap is idempotent, so you can use it to write functions that
take a 'color-ish' argument by passing off the argument to Color.
Currently it takes Colors and hex notation (with or without #).